refactor: Rename tracking.js to analytics.js
