+++ lib/classes/Utils.js
@@ -17,7 +17,7 @@ const walkDirSync = require('../utils/fs/walkDirSync');
 const dirExistsSync = require('../utils/fs/dirExistsSync');
 const isDockerContainer = require('is-docker');
 const version = require('../../package.json').version;
-const tracking = require('../utils/tracking');
+const tracking = require('../utils/analytics');
 const configUtils = require('../utils/config');
 const awsArnRegExs = require('../plugins/aws/utils/arnRegularExpressions');
