+++ lib/classes/Utils.test.js
@@ -8,7 +8,7 @@ const sinon = require('sinon');
 const fse = require('fs-extra');
 const Serverless = require('../../lib/Serverless');
 const configUtils = require('../utils/config');
-const tracking = require('../utils/tracking');
+const tracking = require('../utils/analytics');
 const Utils = require('../../lib/classes/Utils');
 const { expect } = require('chai');
 const { getTmpFilePath, getTmpDirPath } = require('../../tests/utils/fs');
