+++ lib/utils/analytics.test.js
@@ -11,7 +11,7 @@ const cacheDirPath = join(homedir(), '.serverless', 'tracking-cache');
 
 const isFilename = RegExp.prototype.test.bind(/^(?:\.[^.].*|\.\..+|[^.].*)$/);
 
-describe('tracking', () => {
+describe('analytics', () => {
   let track;
   let sendPending;
   let expectedState = 'success';
@@ -34,7 +34,7 @@ describe('tracking', () => {
     return track(type, data);
   };
   before(() => {
-    ({ track, sendPending } = proxyquire('./tracking.js', {
+    ({ track, sendPending } = proxyquire('./analytics.js', {
       './isTrackingDisabled': false,
       'node-fetch': url => {
         usedUrl = url;
