+++ lib/utils/userStats.js
@@ -4,7 +4,7 @@ const BbPromise = require('bluebird');
 const _ = require('lodash');
 const configUtils = require('./config');
 const isValidEventName = require('./userStatsValidation');
-const { track } = require('./tracking');
+const { track } = require('./analytics');
 
 module.exports.track = (eventName, payload) => {
   return BbPromise.try(() => {
